Refactor frontend to native ES modules
- Extracted all classes and logic from monolithic app.js into smaller, maintainable modules in public/src/.
- Created public/src/main.js as the new application entry point.
- Updated index.html to load main.js as a module.
- Organized code by concern: api, map, routing, ui, utils.
- Updated Service Worker to cache all new module files for offline support.
- Updated README.md and AGENTS.md to reflect the new architecture.
- Removed obsolete app.js.

The refactor improves maintainability while keeping the same behavior and no-build deployment model.
