+++ public/index.html
@@ -367,7 +367,7 @@
         crossorigin=""></script>
 
     <!-- App JS -->
-    <script type="module" src="/app.js"></script>
+    <script type="module" src="/src/main.js"></script>
 </body>
 
 </html>
