+++ public/sw.js
@@ -16,7 +16,26 @@ const APP_SHELL = [
     '/',
     '/index.html',
     '/styles.css',
-    '/app.js',
+    '/src/main.js',
+    '/src/config.js',
+    '/src/CircuitWeatherApp.js',
+    '/src/api/F1API.js',
+    '/src/api/WeatherClient.js',
+    '/src/map/MapManager.js',
+    '/src/map/MapWeatherWidget.js',
+    '/src/map/RangeCircles.js',
+    '/src/map/RecentreControl.js',
+    '/src/map/TrackLayer.js',
+    '/src/map/WeatherRadar.js',
+    '/src/routing/Router.js',
+    '/src/ui/CountdownTimer.js',
+    '/src/ui/PrivacyModal.js',
+    '/src/ui/SidebarManager.js',
+    '/src/ui/ThemeManager.js',
+    '/src/ui/WeatherWidget.js',
+    '/src/utils/debounce.js',
+    '/src/utils/escapeHtml.js',
+    '/src/utils/storage.js',
     '/favicon.svg',
     '/manifest.json',
     '/icon-192.png',
